Rename SelectController to SelectHandler and use it like the ShowHandler.
